requirements.txt
flask
__future__
Werkzeug
bs4
importlib
BeautifulSoup
urllib.request
requests
re
statistics
sys
random
time
datetime
googleapiclient.discovery
google_auth_oauthlib.flow
google.auth.transport.requests
functools
operator